Fix so splice() mode does not hang on partial buffers
Before, you might get this:

    $ ( cat BIG_FILE BIG_FILE ) | ./textpv-splice | md5sum
    textpv: 134217728 bytes
    textpv: 287047680 bytes
    textpv: 441843712 bytes
    textpv: 593362944 bytes
    textpv: 748290048 bytes
    textpv: 902692864 bytes
    textpv: 950140928 bytes
    textpv: 950140928 bytes  // same
    textpv: 950140928 bytes  // same
    ...

splice() would block when a partial data transfer was attempted at the
start of the 2nd BIG_FILE.

Instead, we'll allow partial reads, both for the splice() and the read()
case. The only drawback is that we might get small buffers, if the
caller feeds us small chunks: this might result in smallish periodic
reports. But I assume this will not be a common problem.